Close tchannel of a faulty node
When a node gets declared faulty, we want to close the tchannel
connection to it. There are network conditions that a node is not
responding to pings (i.e. is faulty), but the underlying tcp connection
never gets closed. This patch makes it clean the connection for a
faulty.

XXX: currently, tchannel does not provide an easy way to terminate a
connection, so the current implementation is a workaround around it.
Replace this code after there is a nice API to close the tchannel
connection.